Update README.md
In Geant4 and Opticks simulations optical properties like e.g. the materials refractive index, aborption length, rayleigh scattering length, scintillation yield etc.  are inputs that have to be provided by the user. Here we collect the optical properties relevant for liquid Argon TPC's. We collect and decibe the literature resources that were used. In addition we provide scripts that plot this information and can print out gdml stubs that can be included in the gdml detector description.